Exception Handling II

+++ src/main/java/com/project/store/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.project.store.entities.User;
 import com.project.store.repositories.UserRepository;
+import com.project.store.services.exceptions.DatabaseException;
 import com.project.store.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,7 +32,15 @@ public class UserService {
 	}
 	
 	public void deleteUser(Long id) {
-		userRepository.deleteById(id);
+	    try {
+	        if (userRepository.existsById(id)) {
+	            userRepository.deleteById(id);			
+	        } else {				
+	            throw new ResourceNotFoundException(id);			
+	        }		
+	    } catch (DataIntegrityViolationException e) {			
+	        throw new DatabaseException(e.getMessage());		
+	    }	
 	}
 	
 	public User updateUser(Long id, User userUpdated) {
